Split EventLoop tests for promises
